packages/cdk/lambda-backend/requirements.txt
Pillow
boto3
numpy
aws-xray-sdk